refactor(perp): finish deletion of vpool module without variable renames
