+++ x/perp/amm/README.md
@@ -1,0 +1,3 @@
+# `x/perp/amm`
+
+Please refer to our [GitBook](https://docs.nibiru.fi/module-reference/perp/amm).
