--- x/vpool/README.md
@@ -1,3 +1,0 @@
-# `x/vpool`
-
-Please refer to our [GitBook](https://docs.nibiru.fi/module-reference/vpool).
